Moved protected PDFs to subclass
